moving themeProvider render wrapper to testUtils
